requirements.txt
browser-use
playwright
langchain-google-genai
python-dotenv